requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI Integration - PRODUCTION WITH OCR
google-generativeai==0.3.2
pillow==10.2.0
pdf2image==1.17.0
pytesseract==0.3.10
numpy==1.26.4
opencv-python-headless==4.10.0.84

# Elasticsearch
elasticsearch==8.12.0

# Redis
redis==5.0.1
aioredis==2.0.1

# File Processing
PyPDF2==3.0.1

# Email
fastapi-mail==1.4.1

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0